Split out non-public target stuff into new header.
Wrap target methods with functions instead of macros.
